+++ CRM_RK/WebRoot/WEB-INF/customer/updatePc.jsp
@@ -39,9 +39,22 @@ String basePath = request.getScheme()+"://"+request.getServerName()+":"+request.
 		        window.event.returnValue = false;
 		    }
 			else{
-				var idid=$('#cpid').text();
-				window.location.href="<%=path %>/crm/customerPAction_deleteThecustomerp.action?customerP.id="+idid;
-				alert("删除成功");
+				$.ajax({
+					type: "POST",
+					url: "crm/customerPAction_ifcandeleteone.action",
+					data: {deleteid:$('#cpid').text()},
+					dataType: 'text',
+					success: function(result) {
+						if (result=="1") {
+							alert("该客户存在关联数据，请删相关销售机会、订单和服务记录后，再进行此操作");
+							return 
+						}else{
+							var idid=$('#cpid').text();
+							window.location.href="<%=path %>/crm/customerPAction_deleteThecustomerp.action?customerP.id="+idid;
+							alert("删除成功");
+						}
+					}
+				});
 			}				
 		}
 	</script>
